+++ .clang-tidy
@@ -3,11 +3,10 @@ Checks: >
   -*,
   bugprone-*,
   clang-analyzer-*,
-  misc-*,
   modernize-*,
   performance-*,
   portability-*,
-  readability-*,
+  misc-unused-*,
   -bugprone-assignment-in-if-condition,
   -bugprone-branch-clone,
   -bugprone-easily-swappable-parameters,
@@ -16,8 +15,10 @@ Checks: >
   -bugprone-integer-division,
   -bugprone-misplaced-widening-cast,
   -bugprone-narrowing-conversions,
+  -bugprone-reserved-identifier,
   -bugprone-signed-char-misuse,
   -bugprone-string-integer-assignment,
+  -bugprone-stringview-nullptr,
   -bugprone-suspicious-include,
   -bugprone-unchecked-optional-access,
   -clang-analyzer-nullability.NullablePassedToNonnull,
@@ -25,9 +26,6 @@ Checks: >
   -clang-analyzer-optin.osx.*,
   -clang-analyzer-osx.*,
   -clang-analyzer-unix.Malloc,
-  -misc-misplaced-const,
-  -misc-no-recursion,
-  -misc-non-private-member-variables-in-classes,
   -modernize-avoid-c-arrays,
   -modernize-macro-to-enum,
   -modernize-return-braced-init-list,
@@ -35,19 +33,9 @@ Checks: >
   -modernize-use-trailing-return-type,
   -modernize-concat-nested-namespaces,
   -modernize-use-default-member-init,
+  -misc-unused-using-decls,
   -performance-inefficient-string-concatenation,
   -performance-no-int-to-ptr,
-  -readability-braces-around-statements,
-  -readability-container-contains,
-  -readability-convert-member-functions-to-static,
-  -readability-else-after-return,
-  -readability-function-cognitive-complexity,
-  -readability-function-size,
-  -readability-identifier-length,
-  -readability-implicit-bool-conversion,
-  -readability-magic-numbers,
-  -readability-named-parameter,
-  -readability-uppercase-literal-suffix,
 CheckOptions:
   - { key: readability-identifier-naming.ClassCase,           value: CamelCase }
   - { key: readability-identifier-naming.FunctionCase,        value: camelBack }
