+++ STAY/code/common/ecs/type.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <entt/entity/registry.hpp>
+
+namespace stay 
+{
+    namespace ecs
+    {
+        using Entity = entt::entity;
+        using Registry = entt::registry;
+    } // namespace ecs
+} // namespace stay
