+++ STAY/code/common/loader/LDtk/quicktypefwd.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace ldtk 
+{
+    class World;
+    class Level;
+    class Entity;
+    class LayerInstance;
+} // namespace ldtk
