+++ STAY/code/common/loader/sceneLoader.cpp
@@ -1,5 +1,3 @@
-#include <unordered_map>
-#include <filesystem>
 #include <iostream>
 #include <fstream>
 #include <cmath>
