+++ STAY/code/common/physics/layer.hpp
@@ -5,8 +5,6 @@
 #include <cstdint>
 #include <vector>
 
-#include <box2d/box2d.h>
-
 namespace stay
 {
     namespace phys
