+++ STAY/code/common/physics/world.hpp
@@ -14,14 +14,8 @@ namespace stay
             struct WorldTrait
             {
                 using Type = b2World;
-                static Uptr<b2World> create()
-                {
-                    return std::make_unique<b2World>(b2Vec2{0.F, -20.F});
-                }
-                static void shutdown(Uptr<b2World>& world)
-                {
-                    world.reset();
-                }
+                static Uptr<b2World> create();
+                static void shutdown(Uptr<b2World>& world);
             };
         } // namespace details
